### 2017-11-11 #453 @whedon
[PRE REVIEW]: codemetar: Generate CodeMeta Metadata for R Packages
**Submitting author:** @cboettig (<a href="http://orcid.org/0000-0002-1642-628X">Carl Boettiger</a>)
**Repository:** <a href="https://github.com/ropensci/codemetar" target ="_blank">https://github.com/ropensci/codemetar</a>
**Version:** v0.1.1
