### 2017-11-12 #453 @cboettig
Thanks!  Just a note that this package was reviewed by two external reviewers through the rOpenSci onboarding process handled by editor @maelle https://github.com/ropensci/onboarding/issues/130
