### 2017-11-12 #454 @arfon
Accepting because of onboarding into rOpenSci here: https://github.com/ropensci/onboarding/issues/130
