### 2017-11-12 #454 @arfon
@cboettig - could you please merge this PR https://github.com/ropensci/codemetar/pull/39 and once you've made those changes could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
