### 2017-11-13 #454 @cboettig
@arfon Thanks, merged an `paper.bib` added for the links.  Just lemme know if that all looks good and I'll create the release to get the new DOI.
