### 2017-11-13 931d89b
Update NEWS.md
