+++ NEWS.md
@@ -1,3 +1,11 @@
+# codemetar 0.1.2
+
+* JOSS release
+
+# codemetar 0.1.1
+
+* Post onboarding release
+
 # codemetar 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
