### 2017-11-13 #454 @cboettig
@arfon Okay, release made, DOI is: 10.5281/zenodo.1048320
