### 2017-11-13 #454 @arfon
Also, @cboettig - please merge this too when you get a chance https://github.com/ropensci/codemetar/pull/40
