### 2017-11-13 #454 @cboettig
ok, merged, thanks!
